+++ BackEnd/DTOs/UserStockPositionDto.cs
@@ -7,9 +7,9 @@ namespace BackEnd.DTOs
         public string? EnglishName { get; set; }
         public string? Symbol { get; set; }
         public decimal CurrentValue { get; set; }
-        public int BuyPrice { get; set; }
-        public int SellPrice { get; set; }
-        public int ChangeAmount { get; set; }
+        public decimal BuyPrice { get; set; }
+        public decimal SellPrice { get; set; }
+        public decimal ChangeAmount { get; set; }
     }
 }
